FEAT: Create user_id connection for routes validation.

+++ src/Back/model/Items.js
@@ -3,6 +3,7 @@ import { db, PORT } from "../App.js";
 
 export class ItemsModel {
   static async createItem({
+    userID,
     id,
     name,
     description,
@@ -11,6 +12,15 @@ export class ItemsModel {
     sales_price,
     file,
   }) {
+    const ownerCheck = await db.execute({
+      sql: "SELECT id FROM Warehouse WHERE id = ? AND user_id = ?",
+      args: [id, userID],
+    });
+
+    if (ownerCheck.rows.length === 0) {
+      throw new Error("Warehouse not found or access denied");
+    }
+
     const itemId = randomUUID();
     const image_url = file
       ? `http://localhost:${PORT}/uploads/${file.filename}`
@@ -43,6 +53,7 @@ export class ItemsModel {
   }
 
   static async modifyItem({
+    userID,
     warehouseID,
     itemID,
     name,
@@ -57,7 +68,7 @@ export class ItemsModel {
       : null;
 
     const item = await db.execute({
-      sql: `UPDATE Items SET name = ?, description = ?, quantity = ?, purchase_price= ?, sales_price= ?, ${file ? ", image_url = ?" : ""} WHERE id = ? AND warehouse_id = ?`,
+      sql: `UPDATE Items SET name = ?, description = ?, quantity = ?, purchase_price= ?, sales_price = ? ${file ? ", image_url = ?" : ""} WHERE id = ? AND warehouse_id = ? AND warehouse_id IN (SELECT id FROM Warehouse WHERE user_id = ?)`,
       args: file
         ? [
             name,
@@ -68,6 +79,7 @@ export class ItemsModel {
             image_url,
             itemID,
             warehouseID,
+            userID
           ]
         : [
             name,
@@ -77,16 +89,17 @@ export class ItemsModel {
             Number(sales_price),
             itemID,
             warehouseID,
+            userID
           ],
     });
 
     return item.rowsAffected > 0;
   }
 
-  static async deleteItem({ warehouseID, itemID }) {
+  static async deleteItem({ warehouseID, itemID, userID }) {
     const item = await db.execute({
-      sql: "DELETE FROM Items WHERE id = ? AND warehouse_id = ?",
-      args: [itemID, warehouseID],
+      sql: "DELETE FROM Items WHERE id = ? AND warehouse_id = ? AND warehouse_id IN (SELECT id FROM Warehouse WHERE user_id = ?)",
+      args: [itemID, warehouseID, userID],
     });
     return item.rowsAffected > 0;
   }
